requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
facebook-sdk==2.0.0
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
httplib2==0.11.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
psycopg2==2.7.4
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.5
SQLAlchemy-Utils==0.33.1
urllib3==1.22
Werkzeug==0.14.1
